Fixed inconsistent stack_hash() declaration causing compile error on 64-bit.

+++ src/lib/stacktrace.h
@@ -71,7 +71,7 @@ enum stacktrace_sym_quality {
 /**
  * Hashing /equality functions for "struct stacktracea" atomic traces.
  */
-size_t stack_hash(const void *key) G_GNUC_PURE;
+unsigned stack_hash(const void *key) G_GNUC_PURE;
 int stack_eq(const void *a, const void *b) G_GNUC_PURE;
 
 struct logagent;
